pyproject.toml
[tool.poetry]
name = "hexlet-code"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
  { include = "page_analyzer" },
]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^3.0.3"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.1"
flake8 = "^7.0.0"
pytest = "^8.1.1"
psycopg2-binary = "^2.9.9"
validators = "^0.28.3"
requests = "^2.32.3"

[tool.poetry.scripts]
app = "page_analyzer.app:start"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
